Move cloud into pkg/cloud
