+++ cmd/machine-controller/main.go
@@ -24,8 +24,8 @@ import (
 	"k8s.io/apiserver/pkg/util/logs"
 	"sigs.k8s.io/cluster-api/pkg/controller/config"
 
-	"sigs.k8s.io/cluster-api-provider-digitalocean/cloud/digitalocean/controllers/machine"
-	"sigs.k8s.io/cluster-api-provider-digitalocean/cloud/digitalocean/controllers/machine/options"
+	"sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/controllers/machine"
+	"sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/controllers/machine/options"
 )
 
 func main() {
