+++ pkg/cloud/digitalocean/actuators/machine/auth.go
