+++ pkg/cloud/digitalocean/actuators/machine/userdata.go
