+++ pkg/cloud/digitalocean/controllers/machine/controller.go
@@ -14,10 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cluster
+package machine
 
 import (
-	"fmt"
 	"os"
 
 	"github.com/golang/glog"
@@ -34,19 +33,20 @@ import (
 	"k8s.io/client-go/tools/record"
 	"sigs.k8s.io/cluster-api/pkg/client/clientset_generated/clientset"
 	clusterapiclientsetscheme "sigs.k8s.io/cluster-api/pkg/client/clientset_generated/clientset/scheme"
-	"sigs.k8s.io/cluster-api/pkg/controller/cluster"
 	"sigs.k8s.io/cluster-api/pkg/controller/config"
+	machinecontroller "sigs.k8s.io/cluster-api/pkg/controller/machine"
 	"sigs.k8s.io/cluster-api/pkg/controller/sharedinformers"
 
-	clusteractuator "sigs.k8s.io/cluster-api-provider-digitalocean/cloud/digitalocean/actuators/cluster"
-	"sigs.k8s.io/cluster-api-provider-digitalocean/cloud/digitalocean/controllers/cluster/options"
+	machineactuator "sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/actuators/machine"
+	"sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/actuators/machine/machinesetup"
+	"sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/controllers/machine/options"
 )
 
 const (
-	controllerName = "digitalocean-cluster-controller"
+	controllerName = "digitalocean-machine-controller"
 )
 
-func Start(server *options.Server, shutdown <-chan struct{}) {
+func Start(server *options.Server, recorder record.EventRecorder, shutdown <-chan struct{}) {
 	config, err := controller.GetConfig(server.CommonConfig.Kubeconfig)
 	if err != nil {
 		glog.Fatalf("Could not create Config for talking to the apiserver: %v", err)
@@ -57,16 +57,23 @@ func Start(server *options.Server, shutdown <-chan struct{}) {
 		glog.Fatalf("Could not create client for talking to the apiserver: %v", err)
 	}
 
-	params := clusteractuator.ActuatorParams{
-		ClusterClient: client.ClusterV1alpha1().Clusters(corev1.NamespaceDefault),
-	}
-	actuator, err := clusteractuator.NewActuator(params)
+	configWatch, err := machinesetup.NewConfigWatch(server.MachineSetupConfigPath)
 	if err != nil {
-		glog.Fatalf("Could not create digitalocean cluster actuator: %v", err)
+		glog.Fatalf("could not create configWatch: %v", err)
+	}
+
+	params := machineactuator.ActuatorParams{
+		V1Alpha1Client:           client.ClusterV1alpha1(),
+		EventRecorder:            recorder,
+		MachineSetupConfigGetter: configWatch,
+	}
+	actuator, err := machineactuator.NewMachineActuator(params)
+	if err != nil {
+		glog.Fatalf("Could not create digitalocean machine actuator: %v", err)
 	}
 
 	si := sharedinformers.NewSharedInformers(config, shutdown)
-	c := cluster.NewClusterController(config, si, actuator)
+	c := machinecontroller.NewMachineController(config, si, actuator)
 	c.RunAsync(shutdown)
 
 	select {}
@@ -75,24 +82,27 @@ func Start(server *options.Server, shutdown <-chan struct{}) {
 func Run(server *options.Server) error {
 	kubeConfig, err := controller.GetConfig(server.CommonConfig.Kubeconfig)
 	if err != nil {
-		return fmt.Errorf("could not create config for apiserver: %v", err)
+		glog.Errorf("Could not create Config for talking to the apiserver: %v", err)
+		return err
 	}
 
 	kubeClientControl, err := kubernetes.NewForConfig(
-		rest.AddUserAgent(kubeConfig, "cluster-controller-manager"),
+		rest.AddUserAgent(kubeConfig, "machine-controller-manager"),
 	)
 	if err != nil {
-		return fmt.Errorf("invalid API configuration for kubeconfig-control: %v", err)
+		glog.Errorf("Invalid API configuration for kubeconfig-control: %v", err)
+		return err
 	}
 
 	recorder, err := createRecorder(kubeClientControl)
 	if err != nil {
-		return fmt.Errorf("could not create event recorder: %v", err)
+		glog.Errorf("Could not create event recorder : %v", err)
+		return err
 	}
 
 	// run function will block and never return.
 	run := func(stop <-chan struct{}) {
-		Start(server, stop)
+		Start(server, recorder, stop)
 	}
 
 	leaderElectConfig := config.GetLeaderElectionConfig()
@@ -100,13 +110,13 @@ func Run(server *options.Server) error {
 		run(make(<-chan (struct{})))
 	}
 
-	// Identity used to distinguish between multiple cluster controller instances.
+	// Identity used to distinguish between multiple machine controller instances.
 	id, err := os.Hostname()
 	if err != nil {
 		return err
 	}
 
-	leaderElectionClient := kubernetes.NewForConfigOrDie(rest.AddUserAgent(kubeConfig, "cluster-leader-election"))
+	leaderElectionClient := kubernetes.NewForConfigOrDie(rest.AddUserAgent(kubeConfig, "machine-leader-election"))
 
 	id = id + "-" + string(uuid.NewUUID())
 	// Lock required for leader election
@@ -123,7 +133,7 @@ func Run(server *options.Server) error {
 		return err
 	}
 
-	// Try and become the leader and start cluster controller loops
+	// Try and become the leader and start machine controller loops
 	leaderelection.RunOrDie(leaderelection.LeaderElectionConfig{
 		Lock:          rl,
 		LeaseDuration: leaderElectConfig.LeaseDuration.Duration,
