+++ pkg/cloud/digitalocean/controllers/machine/options/options.go
