+++ pkg/cloud/digitalocean/providerconfig/doc.go
@@ -15,8 +15,8 @@ limitations under the License.
 */
 
 // +k8s:deepcopy-gen=package,register
-// +k8s:conversion-gen=sigs.k8s.io/cluster-api-provider-digitalocean/cloud/digitalocean/providerconfig
+// +k8s:conversion-gen=sigs.k8s.io/cluster-api-provider-digitalocean/pkg/cloud/digitalocean/providerconfig
 // +k8s:openapi-gen=true
 // +k8s:defaulter-gen=TypeMeta
 
-package v1alpha1
+package providerconfig
