+++ pkg/cloud/digitalocean/providerconfig/register.go
