+++ pkg/cloud/digitalocean/providerconfig/v1alpha1/types.go
