+++ pkg/cloud/digitalocean/providerconfig/v1alpha1/zz_generated.deepcopy.go
