+++ pkg/cloud/digitalocean/providerconfig/zz_generated.deepcopy.go
